#97-refactor MRG49507
c++20 fixed an argument type.

+++ c++20/mrg49507.h
@@ -115,7 +115,7 @@ public:
     }
 
     /** @brief Valued construtor (integer). */
-    inline Mrg49507(const uint64_t seed) noexcept
+    inline Mrg49507(const uint32_t seed) noexcept
         : MyBaseClass()
     {
         setstate(seed);
